LPS-156398 Moves add attributes method to form provider
