+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/info/item/provider/ObjectEntryInfoItemFormProvider.java
@@ -17,25 +17,42 @@ package com.liferay.object.web.internal.info.item.provider;
 import com.liferay.info.exception.NoSuchFormVariationException;
 import com.liferay.info.field.InfoField;
 import com.liferay.info.field.InfoFieldSet;
+import com.liferay.info.field.type.FileInfoFieldType;
+import com.liferay.info.field.type.NumberInfoFieldType;
+import com.liferay.info.field.type.SelectInfoFieldType;
 import com.liferay.info.form.InfoForm;
 import com.liferay.info.item.field.reader.InfoItemFieldReaderFieldSetProvider;
 import com.liferay.info.item.provider.InfoItemFormProvider;
 import com.liferay.info.localized.InfoLocalizedValue;
+import com.liferay.info.localized.bundle.FunctionInfoLocalizedValue;
+import com.liferay.list.type.model.ListTypeEntry;
+import com.liferay.list.type.service.ListTypeEntryLocalServiceUtil;
+import com.liferay.object.constants.ObjectFieldConstants;
+import com.liferay.object.constants.ObjectFieldValidationConstants;
 import com.liferay.object.exception.NoSuchObjectDefinitionException;
 import com.liferay.object.model.ObjectDefinition;
 import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.model.ObjectField;
+import com.liferay.object.model.ObjectFieldSetting;
 import com.liferay.object.model.ObjectRelationship;
 import com.liferay.object.service.ObjectDefinitionLocalService;
 import com.liferay.object.service.ObjectFieldLocalService;
+import com.liferay.object.service.ObjectFieldSettingLocalServiceUtil;
 import com.liferay.object.service.ObjectRelationshipLocalService;
 import com.liferay.object.web.internal.info.item.ObjectEntryInfoItemFields;
 import com.liferay.object.web.internal.util.ObjectFieldDBTypeUtil;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.template.info.item.provider.TemplateInfoItemFieldSetProvider;
+
+import java.math.BigDecimal;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
 
 /**
  * @author Jorge Ferrer
@@ -107,6 +124,102 @@ public class ObjectEntryInfoItemFormProvider
 		return getInfoForm(formVariationKey);
 	}
 
+	private InfoField<?> _addAttributes(
+		InfoField.FinalStep finalStep, ObjectField objectField) {
+
+		if (Objects.equals(
+				objectField.getBusinessType(),
+				ObjectFieldConstants.BUSINESS_TYPE_ATTACHMENT)) {
+
+			finalStep.attribute(
+				FileInfoFieldType.ALLOWED_FILE_EXTENSIONS,
+				_getAcceptedFileExtensions(objectField)
+			).attribute(
+				FileInfoFieldType.FILE_SOURCE, _getFileSourceType(objectField)
+			).attribute(
+				FileInfoFieldType.MAX_FILE_SIZE,
+				_getMaximumFileSize(objectField)
+			);
+		}
+		else if (Objects.equals(
+					objectField.getBusinessType(),
+					ObjectFieldConstants.BUSINESS_TYPE_DECIMAL)) {
+
+			finalStep.attribute(NumberInfoFieldType.DECIMAL, true);
+		}
+		else if (Objects.equals(
+					objectField.getBusinessType(),
+					ObjectFieldConstants.BUSINESS_TYPE_INTEGER)) {
+
+			finalStep.attribute(
+				NumberInfoFieldType.MAX_VALUE,
+				BigDecimal.valueOf(
+					ObjectFieldValidationConstants.
+						BUSINESS_TYPE_INTEGER_VALUE_MAX)
+			).attribute(
+				NumberInfoFieldType.MIN_VALUE,
+				BigDecimal.valueOf(
+					ObjectFieldValidationConstants.
+						BUSINESS_TYPE_INTEGER_VALUE_MIN)
+			);
+		}
+		else if (Objects.equals(
+					objectField.getBusinessType(),
+					ObjectFieldConstants.BUSINESS_TYPE_LONG_INTEGER)) {
+
+			finalStep.attribute(
+				NumberInfoFieldType.MAX_VALUE,
+				BigDecimal.valueOf(
+					ObjectFieldValidationConstants.BUSINESS_TYPE_LONG_VALUE_MAX)
+			).attribute(
+				NumberInfoFieldType.MIN_VALUE,
+				BigDecimal.valueOf(
+					ObjectFieldValidationConstants.BUSINESS_TYPE_LONG_VALUE_MIN)
+			);
+		}
+		else if (Objects.equals(
+					objectField.getBusinessType(),
+					ObjectFieldConstants.BUSINESS_TYPE_PICKLIST)) {
+
+			finalStep.attribute(
+				SelectInfoFieldType.OPTIONS, _getOptions(objectField));
+		}
+		else if (Objects.equals(
+					objectField.getBusinessType(),
+					ObjectFieldConstants.BUSINESS_TYPE_PRECISION_DECIMAL)) {
+
+			finalStep.attribute(
+				NumberInfoFieldType.DECIMAL, true
+			).attribute(
+				NumberInfoFieldType.DECIMAL_PART_MAX_LENGTH, 16
+			).attribute(
+				NumberInfoFieldType.MAX_VALUE,
+				new BigDecimal(
+					ObjectFieldValidationConstants.
+						BUSINESS_TYPE_PRECISION_DECIMAL_VALUE_MAX)
+			).attribute(
+				NumberInfoFieldType.MIN_VALUE,
+				new BigDecimal(
+					ObjectFieldValidationConstants.
+						BUSINESS_TYPE_PRECISION_DECIMAL_VALUE_MIN)
+			);
+		}
+
+		return finalStep.build();
+	}
+
+	private String _getAcceptedFileExtensions(ObjectField objectField) {
+		ObjectFieldSetting acceptedFileExtensionsObjectFieldSetting =
+			ObjectFieldSettingLocalServiceUtil.fetchObjectFieldSetting(
+				objectField.getObjectFieldId(), "acceptedFileExtensions");
+
+		if (acceptedFileExtensionsObjectFieldSetting == null) {
+			return StringPool.BLANK;
+		}
+
+		return acceptedFileExtensionsObjectFieldSetting.getValue();
+	}
+
 	private InfoFieldSet _getBasicInformationInfoFieldSet() {
 		return InfoFieldSet.builder(
 		).infoFieldSetEntry(
@@ -135,6 +248,31 @@ public class ObjectEntryInfoItemFormProvider
 		).name(
 			"configuration"
 		).build();
+	}
+
+	private FileInfoFieldType.FileSourceType _getFileSourceType(
+		ObjectField objectField) {
+
+		ObjectFieldSetting objectFieldSetting =
+			ObjectFieldSettingLocalServiceUtil.fetchObjectFieldSetting(
+				objectField.getObjectFieldId(), "fileSource");
+
+		if (objectFieldSetting == null) {
+			return null;
+		}
+
+		if (Objects.equals(
+				objectFieldSetting.getValue(), "documentsAndMedia")) {
+
+			return FileInfoFieldType.FileSourceType.DOCUMENTS_AND_MEDIA;
+		}
+		else if (Objects.equals(
+					objectFieldSetting.getValue(), "userComputer")) {
+
+			return FileInfoFieldType.FileSourceType.USER_COMPUTER;
+		}
+
+		return null;
 	}
 
 	private InfoForm _getInfoForm(long objectDefinitionId)
@@ -167,6 +305,18 @@ public class ObjectEntryInfoItemFormProvider
 		).name(
 			_objectDefinition.getClassName()
 		).build();
+	}
+
+	private long _getMaximumFileSize(ObjectField objectField) {
+		ObjectFieldSetting objectFieldSetting =
+			ObjectFieldSettingLocalServiceUtil.fetchObjectFieldSetting(
+				objectField.getObjectFieldId(), "maximumFileSize");
+
+		if (objectFieldSetting == null) {
+			return 0;
+		}
+
+		return GetterUtil.getLong(objectFieldSetting.getValue());
 	}
 
 	private InfoFieldSet _getObjectDefinitionInfoFieldSet(
@@ -214,7 +364,7 @@ public class ObjectEntryInfoItemFormProvider
 					}
 
 					unsafeConsumer.accept(
-						ObjectFieldDBTypeUtil.addAttributes(
+						_addAttributes(
 							InfoField.builder(
 							).infoFieldType(
 								ObjectFieldDBTypeUtil.getInfoFieldType(
@@ -244,6 +394,25 @@ public class ObjectEntryInfoItemFormProvider
 		).name(
 			objectDefinition.getName()
 		).build();
+	}
+
+	private List<SelectInfoFieldType.Option> _getOptions(
+		ObjectField objectField) {
+
+		List<SelectInfoFieldType.Option> options = new ArrayList<>();
+
+		List<ListTypeEntry> listTypeEntries =
+			ListTypeEntryLocalServiceUtil.getListTypeEntries(
+				objectField.getListTypeDefinitionId());
+
+		for (ListTypeEntry listTypeEntry : listTypeEntries) {
+			options.add(
+				new SelectInfoFieldType.Option(
+					new FunctionInfoLocalizedValue<>(listTypeEntry::getName),
+					listTypeEntry.getKey()));
+		}
+
+		return options;
 	}
 
 	private final InfoItemFieldReaderFieldSetProvider
